refactor(pho): abstract document via notecards ((1486))
Original TMX Notebook would output directly to hugo-flavored markdown
documents. We need it now to create abstract documents that SSG adapters
can consume.

Was too hard to read the old code. S-expressions not AST’s. Correct
incorrect names (RSL definition not “footnote”). Smaller functions.

Originally we had the whole rewrite in just over 1000 lines in one
file; but decided it was probably best to keep using at least 2 (and
then 3) files for the main high-level magnetic, and then one file each
for the two largest ancillaries: RSL’s and (much smaller) headers.

Reduce fragility in the RSL def parsing code; now it handles a wider set
of cases and hews more closely to the Gruber spec.

(Still everything is too hacky tho; still this is a proof-of-concept
prototype; because we don’t want to implement about cases like `[a][b]`
(RSL inside inline code span) ever (which this will fall over on.)

We were kind of in a slump. Perhaps needed to take it easy for a week.

(times)
  01-04 09:47  feeling a little stalled … we need to wedge Abstract
               Document in between notecards and the hugo output.
               the code that does footnotes and normalizing header
               levels is kinda messy
  01-11 02:17  now
.