feat(kiss-rdb:eno): introduce caching ((1514))
Discussion: Over the last two-year history of kiss-rdb up to this point,
there was a “KISS” design decision that worked well enough for simple
cases but became an ugliness swept under the rug for more complex
ones: the sharded files of multi-file collections were parsed once per
entity RETRIEVE request. The main philosophy was that a “collection”
is just a “façade”; what you’re really doing usually is figuring out
which one file you need to parse and then scanning it for the one
entity of the request.

For many of our use-cases this approach is fine, but it gets really
ugly when you start using kiss to hold object graphs (a thing we are
doing now and will continue to do to hold progressively larger graphs):
You end up parsing the same storage file multiple times in one request,
one time for every entity you retrieve from it.

We had always known this would be a bridge we would have to cross
eventually. At package birth (2 years ago) we envisioned something
like a caching architecture that would perhaps be adapter agnostic:
a layer that sits between the client and the “backing” storage adapter,
that would itself have the interface of a collection. The idea was you
could insert this access layer between any client and collection to
accomplish caching.

(And yes we wanted to try and use something like redis.)

Now that the time has finally arrived to cross this bridge, we are
realizing that our agnostic idea was a bit too early abstracted; that
what we need has to be tailored to the particularities of the storage
adapter; for example, the eno vendor library parses files as big strings
all at once, etc.

It will be fun and easy.

Also:
  - caching iden strings because it’s time
  - no more overzealous convenience functions in the collection impl.

(times)
  01-16 21:52  begin looking into entity caching
  01-17 12:12  refactored to have “filesystem reader”
        12:47  beging caching
        15:06  refactored existing asset code a second time,
               now using work object “pattern”
        01:46  finished crazy long story about caching & rotating stock
        04:53  now
.