pho(wip): document history for notecard-based documents ((1613))
Journal: Finally seeing this thing come together after 2 or 3 weeks
was thrilling. We cried a few times (joy then terror then joy again):

When it finally game time to ~ bring the two sides of the St. Louis
Arch ~ together, (really it was three sides: 1) populate the data
warehouse 2) use our linguistic processing/bucketing thing, 3)
integrate into the “view” layer (and 3.5 integrate it with pelican
plugin life);

As we tried to stitch these disparate parts together we realized that
we had done zero investigation of what the template looked like near
where the feature was produced that we were going to change.

Initially our target output was “just a paragraph of text (so, lines
of text wordwrapped at no particular place)”; but when we looked at
the structure of the existing HTML we found there was markup and
structure there that we hadn’t planned for:

The page structure had the idea of “created” vs “modified” and also
used locale features of HTML which possibly made their output better
for internationalization.

This sent us “back to the drawing board” a little bit because a big
implicit design vector of our feature was to *remove* structure, and
turn everything into just a stream of words (sentences).

Fortunately (tears of joy again) this wasn’t that painful to retrofit
and regress our pipeline; but it was uncomfortable for a little while.

(times)
  03-06 02:54  begin stats command 
        04:03  stats done! zz
        17:09  ready to integrate looking at template
  03-07 09:34  Don't produce words, produce time bucket expressers
        11:52  lol the first ever real smoke test of the feature and all
               kinds of things are wrong with it. but still, thrilling
        13:44  (it was just one little thing wrong with it in addition
               to another little thing)
               happy with progress, gotta literally run
  03-08 11:06  ~ 1 hour on writing algo for rigged documents
.