feat(pho): document history ((1613))
Discussion: We consider this “feature complete” because it meets our
spec; but true to the spirit of prototyping, now that we see the
behavior in production, we want to make some modifications to the spec.

We want to increase the bucket sizes dynamically so that we target
a target output capacity. Currently for documents with long history,
output is too verbose.

The main work in this commit is of adding support for generating history
for rigged documents. 

(times)
  03-08 12:53  begin designing the schema for rigged document history
               (from the algorithm sketch)
        13:17  ok done writing draft of schema
        13:41  begin looking at writing the warehousing work
               for rigged documents (CLI)
        21:08  resumed. about halfway done with draft of code
  03-09 08:30  code draft and schema changes for populating rigged
               document commits
        15:30  everything is feature complete but now that we see it
               in action, we see we need dynamic buck sizing
        21:30  begin writing commit
.