wip: achieve mostly out-of-the-box CRA
The “tic tac toe” instructions (pointed to in notecard “TL2”) call for
removing all the contents of the “src/” directory immediately after
creating the “create react app” (CRA) starter app.

But we wanted to save a working “snapshot” as soon as possible after
creating the react app with CRA, so, what is in this commit is the
result of the initial CRA command almost exactly:

Because we are choosing to avoid redundant files (with mostly duplicated
content) as much as is reasonable, here we merge the result of CRA
*on top of* what we had working before. (Note this breaks what we
had working before, the part where we use Rollup).

We merged these into one tree so we don't have splintered sub-projects
that get left behind; and we chose to give this (non-Rollup) progression
priority so that we are following the "official" React documentation
as orthodoxy (almost, but not totally exactly, because we still use
that one top-level "package.json" (and now there's a lower one too)).

Also:

  - googled "react live reload not working" because it stopped working
    at some point in the last day. they said it uses "Fast Refresh"
    instead and so see the new option in project.json

(times)
  03-20 15:19  begin reading about tic-tac-toe game before going on run
  03-22 04:48  resume beginnings of tic-tac-toe
        06:00  achieved first step. we can start now
        14:08  got CRA to work with help from stack overflow
  03-23 13:21  so painful, try to continue
        21:58  so painful, try again
  03-24 09:22  please lord show me the light.
               merging almost out-of-the-box CRA
.